Uploaded a new version of the repository.
Uploaded my version of the repository, as requested, due to strange issues folks seem to be having with the previous version where the robots would spin around in circles.  Look at the README beforehand to do any configuring that may need to be taken care of so that it works on your machine.  I would also suggest adding the non-source folders to the gitignore so we aren't committing i.e. logs and other files that don't really have to do with the source code.