Implemented Basic Spiral Search
Implemented a basic spiral search using Fibonacci numbers. For testing, I commented out the code that made the annuli.

Tested with a single rover in a space with no cubes. 

This is a basic implementation that needs improvement. I think the spiral grows outwards too quickly. I might try to implement something with a tigther spiral.