Worked more on spiral search and inner annuli sector
I worked more on spiral search and I think it could still use some improvement? 

I worked on the annuli implementation for the inner sector. One robot is assigned as the organizer and explores within the inner ring (thus staying near the collection zone).

The other two robots are explorers which use the spiral search. At this point they still take cubes all the way back to the collection zone instead of leaving it on the edge of the inner annuli sector.