Reworked spiral search.
Spiral search now chooses waypoints to search using a set of parametric equations, which are independent of the position of the robot.  This will enable the robots to choose points and progress even if, say, a waypoint is chosen in the middle of an obstacle.  This also makes it easy to adjust the tightness of the spiral or where robots start on the spiral.

Also deleted the division of the arena into inner and outer regions, at least for now, due to the fact that the robots behaved in a buggy manner previously.  We can certainly come back to that, but this gives us something we'll be able to test tomorrow in the arena.