Adjusted initial searchpoints and restricted search radius.
Depending on where the robots begin in the arena, the robots will now slightly more reasonably choose points so that they don't collide with one another and drive through the collection zone on startup.  Also, since we'll be searching a 12 m x 12 m square, as opposed to the 15 m x 15 m square in the simulation, adjusted it so that the robot restarts its search if it chooses a point with distance at least 6 m away from the arena center.